+++ examples/react-hybrid-flat/advanced/common/config/main.ts
@@ -6,4 +6,9 @@ export const commonConfig = {
   appName: "react-hybrid-advanced-example",
   version: "1.0.0",
   apiBasePath: "/api",
+  build: {
+    server: {
+      external: ["tailwindcss", "lightningcss"],
+    },
+  },
 };