+++ examples/react-ssr/basic/src/main.ts
@@ -50,6 +50,7 @@ const app = new App({
   build: {
     server: {
       useNativeCompile: false,
+      external: ["tailwindcss", "lightningcss"],
     },
   },
 });