+++ src/utils/runtime.ts
@@ -10,10 +10,7 @@
 import { IS_BUN, IS_DENO, platform } from "@dreamer/runtime-adapter";
 import { DwebErrorCode, throwDwebError } from "./errors.ts";
 
-export {
-  IS_BUN,
-  IS_DENO,
-};
+export { IS_BUN, IS_DENO };
 
 /**
  * 判断当前是否为 Windows 平台